requirements.txt
flask>=3.0.3
youtube-search-python>=1.6.6
yt-dlp>=2024.10.22
pydub>=0.25.1
dropbox>=11.36.2
python-dotenv>=1.0.0